bugfix: fix coordinates ipdetails is null
